Merged in [2248] from ietf@augustcellars.com: The HTML rendering for <xref> elements were inconsistent with the text rendering. Fixed this by doing something completely different than is called for in the bug report:
We follow the layout of what the V3 HTML document says to do.  This
means that we use the child text of the xref when it exists to the
exclusion of any generated text.  When the child text does not exist
then we use the synthesized text string as the text for the anchor
element.  In all cases the anchor element is emitted with an href of the
target.

Uses of the \'cite\' node are removed as they did not seem to be useful.

In a step forward we emit a class on the a node of \'xref\' per the V3
HTML document. Fixes issue #293.
 - Legacy-Id: 2311
Note: SVN reference [2248] has been migrated to Git commit 202793bd4e90e137903d1ad9591a93aef56078a0