Remove pilcrows from tables of contents
In addition to searching list item descendants for the existence of previously-added pilcrows, the code now also searches the list item ancestors for any node that has the 'toc' class to indicate that it is part of a table of contents.  If either are found, the pilcrow is not added.

Fixes #568.  Commit ready for merge.
 - Legacy-Id: 3886